WIP: Rename _version.py to _changelog.py
Make room for automatically generated `_version.py` file by renaming the existing file.
